Final for crops

+++ Cropper/Identification.pde
@@ -10,9 +10,9 @@ class _CropIdentityFactory
 
         // Check for image info
         File base_info_file = new File(identity_folder.getAbsolutePath()+"/"+folder_uuid+"_baseinfo.json");
-        println("Image info path: "+base_info_file);
+        //println("Image info path: "+base_info_file);
         if (base_info_file.exists()) {
-            println("Image info file found successfully");
+            //println("Image info file found successfully");
         } else {
             println("ERROR: No image info file exists for "+folder_uuid);
             return null;
@@ -24,42 +24,33 @@ class _CropIdentityFactory
 
         // Check for imagebase image
         File base_image_file = new File(identity_folder.getAbsolutePath()+"/"+original_image_file_name);
-        println("Base image path: "+base_image_file);
+        //println("Base image path: "+base_image_file);
         if (base_image_file.exists()) {
-            println("Base image file found successfully");
+            //println("Base image file found successfully");
         } else {
             println("ERROR: No base image file exists for "+folder_uuid);
             return null;
         }
 
         File crops_folder = new File(identity_folder.getAbsolutePath()+"/"+"Crops");
-        println("Crop gathering from folder "+crops_folder.getAbsolutePath());
         ArrayList<Crop> crops = new ArrayList<Crop>();
         if (crops_folder.exists()) {
+            println("Crop gathering from folder "+crops_folder.getAbsolutePath());
             File[] crop_files = crops_folder.listFiles();
-            File crop_file;
+            File crop_info_file;
             for (int i = 0; i<crop_files.length; i++) {
-                crop_file = crop_files[i];
-                println("Running processing over crop file: "+crop_file.getAbsolutePath());
-                println("File name: "+crop_file.getName());
-                if (Utils.getExtension(crop_file).equals(OUTPUT_IMAGE_EXTENSION)) {
-                    println("Found Crop file image: "+crop_file.getAbsolutePath());
-                    // subtract 9 from the path to get rid of "_crop.png"
-                    String companion_file_name = crop_file.getAbsolutePath().replace("crop","cropinfo").replace(OUTPUT_IMAGE_EXTENSION, "json");
-                    File crop_info_file = new File(companion_file_name);
-                    println("Searching for companion file: "+crop_info_file);
-                    if (crop_info_file.exists()) {
-                        println("Found compnanion file - adding crop");
-                        crops.add(new Crop(loadJSONObject(crop_info_file.getAbsolutePath())));
-                    } else {
-                        println("ERROR: No companion file found for image "+crop_file.getName());
-                    }
+                crop_info_file = crop_files[i];
+                println("Running processing over crop file: "+crop_info_file.getAbsolutePath());
+                println("File name: "+crop_info_file.getName());
+                if (Utils.getExtension(crop_info_file).equals("json")) {
+                    println("Found cropinfo fileimage: "+crop_info_file.getAbsolutePath());
+                    crops.add(new Crop(loadJSONObject(crop_info_file.getAbsolutePath())));
                 } else {
                     println("Not an image file, continuing...");
                 }
             }
         } else {
-            println("No crop folder existed for image "+image_id);
+            //println("No crop folder existed for image "+image_id);
         }
         return new CropIdentity(base_image_file, image_id, crops);
     }
@@ -90,7 +81,7 @@ class CropIdentity
         } catch (IOException e){
             println("Failed to copy file!"+e);
         }
-        save_to_output_directory();
+        save_info();
     }
 
     // Used for loading. Assumes the original file has already been duplicated.
@@ -103,7 +94,7 @@ class CropIdentity
         this.crops = crops;
     }
 
-    void save_to_output_directory()
+    void save_info()
     {
         JSONObject base_json = new JSONObject();
         base_json.setString("original_image_file_name", original_image_file_name);
@@ -111,24 +102,57 @@ class CropIdentity
         saveJSONObject(base_json, OUTPUT_DIRECTORY+"/"+image_id+"/"+image_id+"_baseinfo.json");
         JSONObject crop_json; 
         
+        delete_old_files("json");
+        
+        for (Crop crop : crops) {
+            crop_json = crop.to_JSON();
+            saveJSONObject(crop_json, OUTPUT_DIRECTORY+"/"+image_id+"/"+"Crops/"+image_id+"_cropinfo_"+crop.crop_index+".json");
+        }
+    }
+    
+    void save_images(){
+        
+        for(int i = 0; i<OUTPUT_IMAGE_EXTENSIONS.length; i++){
+            delete_old_files(OUTPUT_IMAGE_EXTENSIONS[i]);   
+        }
+        
+        for (Crop crop : crops) {
+            crop.crop_image.regenerate_image(loadImage(original_file.getAbsolutePath()));
+            for(int i = 0; i<OUTPUT_IMAGE_EXTENSIONS.length; i++){
+                crop.crop_image.image.save(OUTPUT_DIRECTORY+"/"+image_id+"/"+"Crops/"+image_id+"_crop_"+crop.crop_index+"."+OUTPUT_IMAGE_EXTENSIONS[i]); 
+            }
+        }
+        
+        // tidy up some memory
+        for (Crop crop : crops) {
+            crop.crop_image.image = null;
+        }
+        
+    }
+    
+    void delete_old_files(String extension){
         File crops_directory = new File(OUTPUT_DIRECTORY+"/"+image_id+"/"+"Crops/");
         if(crops_directory.exists()){
             println("Cleaning crops directory "+crops_directory.getAbsolutePath());
             File[] old_crops = crops_directory.listFiles();
             
             for(int i = 0; i<old_crops.length; i++){
-                boolean deletion_success = old_crops[i].delete();
-                println((deletion_success? "Successfully deleted" : "ERROR: Failed to delete") + " old file "+old_crops[i].getAbsolutePath());
+                if(Utils.getExtension(old_crops[i]).equals(extension)){
+                    boolean deletion_success = old_crops[i].delete();
+                    println((deletion_success? "Successfully deleted" : "ERROR: Failed to delete") + " old file "+old_crops[i].getAbsolutePath());
+                }
             }
         }
-        
-        
-        for (Crop crop : crops) {
-            crop_json = crop.to_JSON();
-            saveJSONObject(crop_json, OUTPUT_DIRECTORY+"/"+image_id+"/"+"Crops/"+image_id+"_cropinfo_"+crop.crop_index+".json");
-            crop.crop_image.regenerate_image(loadImage(original_file.getAbsolutePath()));
-            crop.crop_image.image.save(OUTPUT_DIRECTORY+"/"+image_id+"/"+"Crops/"+image_id+"_crop_"+crop.crop_index+"."+OUTPUT_IMAGE_EXTENSION);
+    }
+    
+    String get_crops_json_string(){
+        String ret = "";
+        Crop crop;
+        for (int i = 0; i<crops.size(); i++) {
+            crop = (Crop) crops.get(i);
+            ret += crop.to_JSON().toString();
         }
+        return ret;
     }
 
     String base_image_path()
